refactor: removed separate clock project
- It was only a single function and didn't need to be reusable
