--- sirc-vm/peripheral-clock/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "peripheral_clock"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-spin_sleep_util = "0.1.1"
-spin_sleep = "1.2.0"
-log = "0.4.21"
