+++ sirc-vm/sbrc-vm/Cargo.toml
@@ -12,7 +12,6 @@ device_terminal = { path = "../device-terminal" }
 device_video = { path = "../device-video" }
 log = "0.4.21"
 peripheral_bus = { path = "../peripheral-bus" }
-peripheral_clock = { path = "../peripheral-clock" }
 peripheral_cpu = { path = "../peripheral-cpu" }
 stderrlog = "0.6.0"
 dap = "0.4.1-alpha1"
@@ -20,6 +19,8 @@ thiserror = "1.0.61"
 postcard = { version = "1.0.8", features = ["alloc"] }
 serde = "1.0.200"
 line-col = "0.2.1"
+spin_sleep_util = "0.1.1"
+spin_sleep = "1.2.0"
 
 [dev-dependencies]
 criterion = { version = "0.5", features = ["html_reports"] }
