+++ sirc-vm/sbrc-vm/src/utils/mod.rs
@@ -1,2 +1,3 @@
+pub mod frame_reporter;
 pub mod lines;
 pub mod path;
